Switch UserRequest constructor to take a double instead of a float

+++ src/main/java/com/kiafarhang/distillr/server/UserRequest.java
@@ -15,7 +15,7 @@ public class UserRequest {
         this.searchTerm = searchTerm;
     }
 
-    public UserRequest(Location location, float money, String searchTerm) {
+    public UserRequest(Location location, double money, String searchTerm) {
         this.location = location;
         this.money = money;
         this.searchTerm = searchTerm;
